http-rpc-common module is now rest-common
