+++ components/flow/flow-rpcops-service-impl/build.gradle
@@ -17,7 +17,7 @@ dependencies {
     implementation project(':components:permissions:permission-validation-service')
     implementation project(':libs:configuration:configuration-core')
     implementation project(':libs:http-rpc:http-rpc')
-    implementation project(':libs:http-rpc:http-rpc-common')
+    implementation project(':libs:http-rpc:rest-common')
     implementation project(':libs:lifecycle:lifecycle')
     implementation project(':libs:messaging:messaging')
     implementation project(':libs:virtual-node:virtual-node-info')
