+++ components/permissions/permission-rpc-ops-impl/build.gradle
@@ -19,7 +19,7 @@ dependencies {
     implementation project(':libs:permissions:permission-endpoint')
     implementation project(':libs:permissions:permission-manager')
     implementation project(':libs:permissions:permission-common')
-    implementation project(":libs:http-rpc:http-rpc-common")
+    implementation project(":libs:http-rpc:rest-common")
     implementation project(":libs:http-rpc:http-rpc")
     implementation project(":libs:utilities")
