+++ components/virtual-node/virtual-node-rpcops-maintenance-impl/build.gradle
@@ -16,7 +16,7 @@ dependencies {
 
     implementation project(':libs:virtual-node:virtual-node-endpoints-maintenance')
     implementation project(':libs:http-rpc:http-rpc')
-    implementation project(':libs:http-rpc:http-rpc-common')
+    implementation project(':libs:http-rpc:rest-common')
     implementation project(':libs:lifecycle:lifecycle')
     implementation project(':libs:messaging:messaging')
     implementation project(":libs:utilities")
