+++ components/virtual-node/virtual-node-rpcops-maintenance-impl/src/test/kotlin/net/corda/libs/virtualnode/maintenance/rpcops/impl/v1/VirtualNodeMaintenanceRPCOpsImplTest.kt
@@ -3,8 +3,8 @@ package net.corda.libs.virtualnode.maintenance.rpcops.impl.v1
 import net.corda.chunking.ChunkWriter
 import net.corda.cpi.upload.endpoints.service.CpiUploadRPCOpsService
 import net.corda.httprpc.HttpFileUpload
-import net.corda.httprpc.security.CURRENT_RPC_CONTEXT
-import net.corda.httprpc.security.RpcAuthContext
+import net.corda.httprpc.security.CURRENT_REST_CONTEXT
+import net.corda.httprpc.security.RestAuthContext
 import net.corda.libs.cpiupload.CpiUploadManager
 import net.corda.lifecycle.LifecycleCoordinator
 import net.corda.lifecycle.LifecycleCoordinatorFactory
@@ -30,10 +30,10 @@ class VirtualNodeMaintenanceRPCOpsImplTest {
         @JvmStatic
         @BeforeAll
         fun setRPCContext() {
-            val rpcAuthContext = mock<RpcAuthContext>().apply {
+            val restAuthContext = mock<RestAuthContext>().apply {
                 whenever(principal).thenReturn(actor)
             }
-            CURRENT_RPC_CONTEXT.set(rpcAuthContext)
+            CURRENT_REST_CONTEXT.set(restAuthContext)
         }
     }
     private val mockCoordinator = mock<LifecycleCoordinator>().apply {
