+++ components/virtual-node/virtual-node-rpcops-service-impl/build.gradle
@@ -18,7 +18,7 @@ dependencies {
     implementation project(':libs:virtual-node:virtual-node-endpoints')
     implementation project(':libs:virtual-node:virtual-node-info')
     implementation project(':libs:http-rpc:http-rpc')
-    implementation project(':libs:http-rpc:http-rpc-common')
+    implementation project(':libs:http-rpc:rest-common')
     implementation project(':libs:lifecycle:lifecycle')
     implementation project(':libs:messaging:messaging')
     implementation project(":libs:utilities")
