+++ components/virtual-node/virtual-node-rpcops-service-impl/src/main/kotlin/net/corda/virtualnode/rpcops/impl/v1/VirtualNodeRestResourceImpl.kt
@@ -15,7 +15,7 @@ import net.corda.httprpc.PluggableRestResource
 import net.corda.httprpc.exception.InternalServerException
 import net.corda.httprpc.exception.InvalidInputDataException
 import net.corda.httprpc.exception.ResourceNotFoundException
-import net.corda.httprpc.security.CURRENT_RPC_CONTEXT
+import net.corda.httprpc.security.CURRENT_REST_CONTEXT
 import net.corda.libs.configuration.helper.getConfig
 import net.corda.libs.cpiupload.endpoints.v1.CpiIdentifier
 import net.corda.libs.virtualnode.endpoints.v1.VirtualNodeRestResource
@@ -210,7 +210,7 @@ internal class VirtualNodeRestResourceImpl @Activate constructor(
         )
         validateX500Name(request.x500Name)
 
-        val actor = CURRENT_RPC_CONTEXT.get().principal
+        val actor = CURRENT_REST_CONTEXT.get().principal
         val rpcRequest = with(request) {
             VirtualNodeManagementRequest(
                 instant,
@@ -263,7 +263,7 @@ internal class VirtualNodeRestResourceImpl @Activate constructor(
     ): ChangeVirtualNodeStateResponse {
         val instant = clock.instant()
         // Lookup actor to keep track of which RPC user triggered an update
-        val actor = CURRENT_RPC_CONTEXT.get().principal
+        val actor = CURRENT_REST_CONTEXT.get().principal
         logger.debug { "Received request to update state for $virtualNodeShortId to $newState by $actor at $instant" }
         if (!isRunning) throw IllegalStateException(
             "${this.javaClass.simpleName} is not running! Its status is: ${lifecycleCoordinator.status}"
