+++ components/virtual-node/virtual-node-rpcops-service-impl/src/test/kotlin/net/corda/configuration/rpcops/impl/v1/VirtualNodeRPCOpsImplTest.kt
@@ -1,7 +1,7 @@
 package net.corda.configuration.rpcops.impl.v1
 
-import net.corda.httprpc.security.CURRENT_RPC_CONTEXT
-import net.corda.httprpc.security.RpcAuthContext
+import net.corda.httprpc.security.CURRENT_REST_CONTEXT
+import net.corda.httprpc.security.RestAuthContext
 import net.corda.lifecycle.LifecycleCoordinator
 import net.corda.lifecycle.LifecycleCoordinatorFactory
 import net.corda.utilities.time.ClockFactory
@@ -26,10 +26,10 @@ class VirtualNodeRPCOpsImplTest {
         @JvmStatic
         @BeforeAll
         fun setRPCContext() {
-            val rpcAuthContext = mock<RpcAuthContext>().apply {
+            val restAuthContext = mock<RestAuthContext>().apply {
                 whenever(principal).thenReturn(actor)
             }
-            CURRENT_RPC_CONTEXT.set(rpcAuthContext)
+            CURRENT_REST_CONTEXT.set(restAuthContext)
         }
     }
