--- libs/http-rpc/http-rpc-common/README.md
@@ -1,3 +1,0 @@
-# Corda HTTP RPC Common
-
-This module contains classes required by other HTTP RPC modules to implement both server and client side functionality.
