+++ libs/http-rpc/http-rpc-server-impl/build.gradle
@@ -45,7 +45,7 @@ dependencies {
     }
 
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
-    implementation project(":libs:http-rpc:http-rpc-common")
+    implementation project(":libs:http-rpc:rest-common")
 
     implementation "com.fasterxml.jackson.datatype:jackson-datatype-jsr310:$jacksonVersion"
