+++ libs/http-rpc/http-rpc-stubs-impl/build.gradle
@@ -14,7 +14,7 @@ dependencies {
     implementation project(":libs:http-rpc:http-rpc")
     api project(":libs:http-rpc:ssl-cert-read")
     implementation project(":libs:http-rpc:http-rpc-security-read")
-    implementation project(":libs:http-rpc:http-rpc-common")
+    implementation project(":libs:http-rpc:rest-common")
     implementation project(":libs:utilities")
 
     implementation "org.apache.commons:commons-lang3:$commonsLangVersion"
