+++ libs/http-rpc/http-rpc-test-common/build.gradle
@@ -19,7 +19,7 @@ dependencies {
     implementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
 
     implementation project(":libs:http-rpc:http-rpc")
-    implementation project(":libs:http-rpc:http-rpc-common")
+    implementation project(":libs:http-rpc:rest-common")
     implementation project(":libs:http-rpc:rest-tools")
     implementation project(":libs:http-rpc:http-rpc-security-read")
     implementation project(":libs:serialization:serialization-amqp")
