+++ libs/http-rpc/rest-common/README.md
@@ -1,0 +1,3 @@
+# Corda Rest Common
+
+This module contains classes required by other Rest modules to implement both server and client side functionality.
