+++ libs/http-rpc/rest-common/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'corda.common-publishing'
     id 'corda.common-library'
 }
-description 'Corda HTTP RPC Common'
+description 'Corda Rest Common'
 
 dependencies {
