+++ libs/http-rpc/rest-common/detekt-baseline.xml
