+++ libs/http-rpc/rest-common/src/main/kotlin/net/corda/httprpc/durablestream/api/FiniteDurableCursorBuilder.kt
