+++ libs/http-rpc/rest-common/src/test/java/net/corda/httprpc/durablestream/api/CursorExceptionJavaApiTest.java
