+++ libs/http-rpc/rest-tools/build.gradle
@@ -11,7 +11,7 @@ dependencies {
     implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
     implementation "net.corda:corda-base"
     implementation project(':libs:http-rpc:http-rpc')
-    implementation project(':libs:http-rpc:http-rpc-common')
+    implementation project(':libs:http-rpc:rest-common')
 
     compileOnly "org.osgi:osgi.annotation"
 }
