+++ settings.gradle
@@ -247,7 +247,7 @@ include 'libs:http-rpc:http-rpc-server'
 include 'libs:http-rpc:http-rpc-server-impl'
 include 'libs:http-rpc:http-rpc-stubs-impl'
 include 'libs:http-rpc:ssl-cert-read'
-include 'libs:http-rpc:http-rpc-common'
+include 'libs:http-rpc:rest-common'
 include 'libs:http-rpc:rest-client'
 include 'libs:http-rpc:http-rpc-test-common'
 include 'libs:http-rpc:rbac-security-manager'
